third_party/WebKit/LayoutTests/svg/dom/SVGScriptElement/script-external-no-multiple-load.html

<!DOCTYPE html>
<script src=../../../resources/js-test.js></script>
<script>
window.jsTestIsAsync = true;

var loadEventsFired = 0;

onload = function() {
    var newExternalScript = document.createElementNS("http://www.w3.org/2000/svg", "script");
    newExternalScript.onload = function() {
        loadEventsFired++;
        setTimeout(function() {
            shouldBe("loadEventsFired", "1");
            finishJSTest();
        }, 0);
    };
    newExternalScript.setAttributeNS("http://www.w3.org/1999/xlink", "href", "resources/script-load.js");
    document.querySelector("svg").appendChild(newExternalScript);
}
</script>
<svg>
</svg>
