third_party/blink/web_tests/svg/dom/SVGScriptElement/script-external-no-multiple-load-expected.txt

PASS loadEventsFired is 1
PASS successfullyParsed is true

TEST COMPLETE
